--- .prettierrc
@@ -1,22 +1,0 @@
-{
-  "printWidth": 80,
-  "tabWidth": 2,
-  "useTabs": false,
-  "semi": true,
-  "singleQuote": false,
-  "jsxSingleQuote": false,
-  "jsxBracketSameLine": false,
-  "trailingComma": "none",
-  "singleAttributePerLine": true,
-  "bracketSpacing": true,
-  "arrowParens": "avoid",
-  "proseWrap": "always",
-  "importOrder": ["^react$", "^next", "<THIRD_PARTY_MODULES>", "^@/", "^[./]"],
-  "importOrderSeparation": true,
-  "importOrderSortSpecifiers": true,
-  "importOrderCaseInsensitive": true,
-  "plugins": [
-    "@trivago/prettier-plugin-sort-imports",
-    "prettier-plugin-tailwindcss"
-  ]
-}
