--- agent/README.md
@@ -1,15 +1,0 @@
-# Dexter.ai Security Standards Assistant
-
-An AI-powered assistant for security standards and compliance guidance.
-
-## Installation
-
-```bash
-uv pip install -e .
-```
-
-## Usage
-
-```bash
-python main.py
-```
