--- server/eslint.config.js
@@ -1,37 +1,0 @@
-import tsPlugin from "@typescript-eslint/eslint-plugin";
-import tsParser from "@typescript-eslint/parser";
-import prettierConfig from "eslint-config-prettier";
-import prettierPlugin from "eslint-plugin-prettier";
-
-export default [
-  {
-    ignores: ["node_modules/**", "dist/**"],
-  },
-  {
-    files: ["**/*.ts"],
-    languageOptions: {
-      parser: tsParser,
-      parserOptions: {
-        project: "./tsconfig.json",
-        ecmaVersion: "latest",
-        sourceType: "module",
-      },
-    },
-    plugins: {
-      "@typescript-eslint": tsPlugin,
-      prettier: prettierPlugin,
-    },
-    rules: {
-      ...tsPlugin.configs.recommended.rules,
-      ...prettierPlugin.configs.recommended.rules,
-      "prettier/prettier": "error",
-      "@typescript-eslint/no-unused-vars": "warn",
-      "@typescript-eslint/no-explicit-any": "warn",
-      "@typescript-eslint/explicit-function-return-type": "warn",
-      "@typescript-eslint/explicit-module-boundary-types": "warn",
-      "@typescript-eslint/no-floating-promises": "error",
-      "no-console": ["warn", { allow: ["warn", "error"] }],
-    },
-  },
-  prettierConfig,
-];
