+++ server/tsconfig.json
@@ -5,8 +5,18 @@
     "strict": true,
     "verbatimModuleSyntax": true,
     "skipLibCheck": true,
+    "noEmit": true,
     "types": ["node"],
     "jsx": "react-jsx",
-    "jsxImportSource": "hono/jsx"
-  }
+    "jsxImportSource": "hono/jsx",
+    "baseUrl": ".",
+    "outDir": "./dist",
+    "allowJs": true,
+    "checkJs": true,
+    "paths": {
+      "@/*": ["src/*"]
+    }
+  },
+  "include": ["eslint.config.mjs", "prettier.config.js", "**/*.ts", "**/*.js"],
+  "exclude": ["node_modules", "dist"]
 }
