--- web/drizzle.config.ts
@@ -1,12 +1,0 @@
-import { type Config } from "drizzle-kit";
-
-import { env } from "@/env";
-
-export default {
-  schema: "./src/server/db/schema.ts",
-  dialect: "sqlite",
-  dbCredentials: {
-    url: env.DATABASE_URL,
-  },
-  tablesFilter: ["web_*"],
-} satisfies Config;
