+++ web/next.config.js
@@ -5,14 +5,6 @@
 import "./src/env.js";
 
 /** @type {import("next").NextConfig} */
-const config = {
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  output: "standalone",
-};
+const config = {};
 
 export default config;
