+++ web/postcss.config.js
@@ -1,5 +1,7 @@
-export default {
+const config = {
   plugins: {
-    tailwindcss: {},
-  },
+    "@tailwindcss/postcss": {}
+  }
 };
+
+export default config;
